requirements.txt
numpy
tensorflow
matplotlib
opencv-contrib-python
pillow
